restructured project to match conventional web frameworks like spring, rails etc
