+++ src/lib.rs
@@ -1,9 +1,10 @@
 pub mod models;
-pub mod schema;
-pub mod oauth;
 pub mod db;
 pub mod state;
+pub mod routes;
 pub mod config;
+pub mod schema;
+pub mod oauth;
 
 #[macro_use]
 extern crate serde_derive;
