+++ src/models.rs
@@ -1,16 +1,1 @@
-use super::schema::users;
-
-#[derive(Serialize, Queryable)]
-pub struct User {
-    pub id: String,
-    pub name: String,
-    pub password: String,
-}
-
-#[derive(Insertable)]
-#[table_name = "users"]
-pub struct NewUser<'a> {
-    pub id: &'a str,
-    pub name: &'a str,
-    pub password: &'a str,
-}
+pub mod user;
