+++ src/routes.rs
@@ -1,0 +1,2 @@
+pub mod oauth;
+pub mod user;
